# openapi (4).yaml
openapi: 3.0.3
info:
  title: Swagger Travel Agency - OpenAPI 3.0

  
  version: 1.0.11
servers:
  - url: https://agency.swagger.io/api/v3
...
paths:
  /tour:
    put:
      tags:
        - tour
      summary: Update an existing tour
      description: Update an existing tour by Id
      operationId: updateTour
      requestBody:
        description: Update an existent tour in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tour'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tour'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tour'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Tour'
        '400':
          description: Invalid ID supplied
        '404':
          description: Tour not found
        '405':
          description: Validation exception
      security:
        - agency_auth:
            - write:tours
            - read:tours
    post:
      tags:
        - tour
      summary: Add a new tour to the agency
      description: Add a new tour to the agency
      operationId: addTour
      requestBody:
        description: Create a new tour in the agency
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTour'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewTour'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewTour'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Tour'
            
        '405':
          description: Invalid input
      security:
        - agency_auth:
            - write:tours
            - read:tours
...
  /order:
    post:
      tags:
        - order
      summary: Place an order for a tour
      description: Place a new order in the agency
      operationId: placeOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrder'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewOrder'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewOrder'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      #   '405':
      #     description: Invalid input
      security:
        - api_key: []
...
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewUser'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Empty fields
        '401':
          description: Username or email already taken
...
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        
        quantity:
          type: integer
          format: int32
          example: 7
        orderDate:
          type: string
          format: date-time
        status:
          type: string
          description: Order Status
          example: approved
          enum:
            - placed
            - approved
            - recived
        complete:
          type: boolean
        idTour:
            $ref: '#/components/schemas/Tour'
        idUser:
            $ref: '#/components/schemas/User'
      xml:
        name: order
    NewOrder:
      type: object
      properties:
        
        quantity:
          type: integer
          format: int32
          example: 7
        orderDate:
          type: string
          format: date-time
        status:
          type: string
          description: Order Status
          example: approved
          enum:
            - placed
            - approved
            - recived
        complete:
          type: boolean
        idTour:
            $ref: '#/components/schemas/Tour'
        idUser:
            $ref: '#/components/schemas/User'
      xml:
        name: neworder
    User:
      type: object
      properties:
        role:
         type: string
         enum:
            - admin
            - customer
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
         
      xml:
        name: user
    NewUser:
      type: object
      properties:
        role:
         type: string
         enum:
            - admin
            - customer
        
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
         
      xml:
        name: newuser
    Tour:
      required:
        - name
        - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: New Zeland tour
        value:
          type: integer
          example: 400
        photoUrls:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            xml:
              name: photoUrl
        status:
          type: string
          description: tour status in the agency
          enum:
            - available
            - sold
        
      xml:
        name: tour
    NewTour:
      required:
        - name
        - photoUrls
      type: object
      properties:
       
        name:
          type: string
          example: New Zeland tour
        value:
          type: integer
          example: 400
        photoUrls:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            xml:
              name: photoUrl
        status:
          type: string
          description: tour status in the agency
          enum:
            - available
            - sold
      xml:
        name: newtour   
...
  securitySchemes:
    agency_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://agency.swagger.io/oauth/authorize
          scopes:
            write:tours: modify tours in your account
            read:tours: read your tours
    api_key:
      type: apiKey
      name: api_key
      in: header